Hoist filter load if alpha == beta == 0 in warp_affine_neon
In warp affine the common case is that both alpha and beta are zero, in
which case the vector of filter values does not need to be re-loaded on
each iteration of the loop.

LLVM 18 appears to struggle to do this itself, so manually hoist out the
vector and pass it as a parameter instead.

On a Neoverse V2 machine this seems to give a geomean ~0.9% reduction in
times measured for the warp affine speed tests for the cases where alpha
and beta are both zero.

This is a port of libaom commit
538a49e696266328264ff694da3b6f302befce68.

Also put some helper functions in a separate header file, as they will
be used in I8MM and SVE implementations of warp_affine.