Optimize Neon implementation of svt_aom_lpf_14
The Neon implementations of svt_aom_lpf_vertical_14 and
svt_aom_lpf_horizontal_14 compute every filter (filter4(), filter8() and
filter14()) before selecting for each element which filter is actually
needed. In practice, however, a lot of cases only need one of the
filters, so specialize for these scenarios, computing only the filters
that are needed and eliminate bitwise select. This makes the case where
all filters are needed slightly slower, but as it is far from the most
common case this is ok.

Also move the actual filter computation to separate functions to avoid
code duplication.