Refactor unit tests for svt_av1_calc_target_weighted_pred
Split the test instantiation to use the same naming convention as other
unit tests, and use ASSERT_EQ to check for equality, to avoid errors
being printed but the tests being shown as PASSED.

Also pass the different overlap values directly in the test
instantiation, which allows directly printing the value in case one test
fails.

Finally, remove the test case where overlap = 1, as overlap is always
half of a block dimension, and the smallest possible dimension is 4.